\name{set.wpp.year}
\alias{set.wpp.year}
\alias{get.wpp.year}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Edit/Retrieve WPP Year.
}
\description{
Change/get the WPP year which determines from which \pkg{wpp} package data are retrieved.
}
\usage{
set.wpp.year(wpp.year, sim.dir = NULL)

get.wpp.year()
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{wpp.year}{Currently only 2015 is available.}
  \item{sim.dir}{Simulation directory used for probability calculations.}
}
\details{
	The package operates on the \pkg{wpp2015BayesMig} data package only. Using the \code{set.wpp.year} has the side effect of clearing up used data from the memory. It can be used to change the simulation directory which has to be a directory generated by the \pkg{bayesPop} package. 
}
\value{
\code{get.wpp.year} returns the wpp year used through out the package.
}


\examples{
get.wpp.year() # by default the wpp2015 is used
}
% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ manip }

